Correctly handle end of episodes with ValueStructureKernels: the first step of the new episode is NOT considered a successor of the last step of the previous episode anymore

+++ edge/model/inference/tensorwrap.py
@@ -3,9 +3,9 @@ from inspect import signature
 from decorator import decorator
 
 
-def ensure_tensor(x):
+def ensure_tensor(x, dtype=torch.float):
     if not torch.is_tensor(x):
-        x = torch.tensor(x).float()
+        x = torch.tensor(x).type(dtype)
     return x
